Several fixes for Windows. (st)
git-svn-id: svn://mielke.cc/main/brltty@1739 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ ScreenDrivers/Windows/screen.c
@@ -54,7 +54,7 @@ close_WindowsScreen (void) {
 static void
 describe_WindowsScreen (ScreenDescription *description) {
   CONSOLE_SCREEN_BUFFER_INFO info;
-  description->no = GetForegroundWindow();
+  description->no = (int) GetForegroundWindow();
   if (consoleOutput == INVALID_HANDLE_VALUE)
     return;
   if (!(GetConsoleScreenBufferInfo(consoleOutput, &info)))
